+++ mopidy/stream/actor.py
@@ -8,7 +8,7 @@ import urlparse
 import pykka
 
 from mopidy import audio as audio_lib, backend, exceptions
-from mopidy.audio import scan
+from mopidy.audio import scan, utils
 from mopidy.models import Track
 
 logger = logging.getLogger(__name__)
@@ -31,7 +31,7 @@ class StreamBackend(pykka.ThreadingActor, backend.Backend):
 class StreamLibraryProvider(backend.LibraryProvider):
     def __init__(self, backend, timeout, blacklist):
         super(StreamLibraryProvider, self).__init__(backend)
-        self._scanner = scan.Scanner(min_duration=None, timeout=timeout)
+        self._scanner = scan.Scanner(timeout=timeout)
         self._blacklist_re = re.compile(
             r'^(%s)$' % '|'.join(fnmatch.translate(u) for u in blacklist))
 
@@ -44,8 +44,9 @@ class StreamLibraryProvider(backend.LibraryProvider):
             return [Track(uri=uri)]
 
         try:
-            data = self._scanner.scan(uri)
-            track = scan.audio_data_to_track(data)
+            tags, duration = self._scanner.scan(uri)
+            track = utils.convert_tags_to_track(tags).copy(
+                uri=uri, length=duration)
         except exceptions.ScannerError as e:
             logger.warning('Problem looking up %s: %s', uri, e)
             track = Track(uri=uri)