+++ tests/audio/test_listener.py
@@ -8,7 +8,7 @@ from mopidy import audio
 
 
 class AudioListenerTest(unittest.TestCase):
-    def setUp(self):
+    def setUp(self):  # noqa
         self.listener = audio.AudioListener()
 
     def test_on_event_forwards_to_specific_handler(self):
@@ -32,3 +32,6 @@ class AudioListenerTest(unittest.TestCase):
 
     def test_listener_has_default_impl_for_position_changed(self):
         self.listener.position_changed(None)
+
+    def test_listener_has_default_impl_for_tags_changed(self):
+        self.listener.tags_changed([])